pushing GVR and Unity 5.6 updated code...
Modified script for splineController to use enum, easier to change spline state using other scripts.  Working on properly starting and stopping spline based upon triggers

Had some issues with starting the spline, partially due to weird changes with the scaling of my playerContainer and the GVR 1.5 update.  Also, am now seeing two reticles and can't move the player camera using the keyboard…
-Upgraded to 5.6.1 of unity, reinstalled GVR SDK, and changed platform.  The camera movement is working now.  Not sure about the multi reticles

-Spent an hour trying to figure out why the reticle was not responding to my interactive objects properly!  Ending up just importing googles "DemoInputManager" prefab and now the reticle is working.

Issues & Fixes with player movement separate from the spline movement.  Reorganized the game object structure of the mainCamera and parent objects.  Set the playerCube as a parent of the mainCamera and set the player character a child of the mainCamera.
-Also adjusted the script to only rotate the player character object with the camera rotation.  But move the playerCube in the direction of the main camera
-constrained rotation of playerCube to never rotate (don't want it to rotate, just the camera and player character.